Fix sizing bug
When the grid's dimension was reduced to below 16 x 16
the space wasn't distributed evenly, and there was some
space  leftin the grid,, which the squares weren't occupying.

I fixed the bug by adding 'Flex: 1' to the .square class, so
that all the grid would be filled, and all the squares would
have an equal size; the other changes weren't relevant to this,
and were just some adjustments i thought of while debugging.