Fix eventListener bug
When the grid changed dimensions (a new one was created), the
'click', and 'mouseenter' events weren't being triggered. So i
amoved the addition of the events to right as the squares are
being created (in the createSquares function).

As bonus, this was an optimization, since now we wont't have to
go through the whole squares array/nodeList to perform the
operation. We are there when the nodeList is being created.

The change of events from mousemove, to mouseenter has nothing to
do with above. I just changed because it feels smoother, and also
performs less operations.